Creates AnsibleModuleHelper
